Only set `quarkus.http.host` as a default if not available
